Allow sizeof to be used in constant expressions.
Add proper support for the sizeof operator so that both "sizeof a" and
"sizeof(a)" is supported.

This involved expanding parse_op to recognize a "special" operator
which is just an operator, see ops.in.  In the process I removed the
Group class since it wasn't needed and just got in the way.

svn-id: zl/trunk@52