Check point.
Add primitive support for classes.  No inheritance, just objects with
methods.  Still some issues to work out, in particular class members
can not be used as a rvalue.  Also "this->foo" not handles correctly.

Modules are now parsed using two passes.  The first pass parses
everything except function and variable bodies.  The second pass
parses the bodies.

Allow macros to be used in identifier position.  At the moment there is
no way to distinguish when a macro is being called as a function with
no parameters and when it is being called as an identifier.  Will fix
latter.

Enhance test script to track failed tests.

Convert Syntax (and friends) print methods to to_string(OStream &)

svn-id: zl/trunk@115