Added symbol_exists callback function.
Fixed grammar of "for".

Don't cast initialize lists as that causes C to think it is no longer
a constant.  Need to fix to resolve individual components of list
instead.

Deref ref types when passed in to functions with "...", and add hook
to eventually perform correct "default argument promotion" as
specified in the standard.

Change test61 to match how it is defined the submitted gpce09 paper
and add test for other gpcde9 paper examples.

svn-id: trunk/zl@319