More refactoring, for better or worse.
Fix things up so only one instance of Empty (now EmptyStmt) needs to exist.

Declaration now a direct descendant of AST rather than STMT.  Var is a
descendant of both Stmt and Declaration.  This means that AST has to
be a virtual base, yuck.

Finalize is now called directly by Fun, since that is the only place
it used.

svn-id: trunk/zl@348