Checkpoint.  For better or worse...
Get rid of separate top level symbol list.  They are not just part of
the environ.  For example static variables inside a function are added
twice one to the local environment in which they where declared and once
to the top level environ.  The first is just an alias pointing to the
latter. The latter symbol key is the managed name (called uniq_name).
Some symbols are also now marked as internal to keep them from being
outputted during compilation.

Many other small changes due to this.  Including having to use another
instance of virtual inheritance (uck!).

svn-id: trunk/zl@367