Checkpoint towards being able to print fully expanded output.
When compiling for expanded output gather marks and repl-tables to
avoid printing the same object twice.  Also, when compiling parts that
need this compile to a string rather than to a file.  The marks and
repl-tables are then printed first with the rest coming after.

To do this a new class is created SyntaxGather with a pointer in the
CompileWriter when it is needed.  Syntax related to_string functions
now take it as an optional argument.

svn-id: trunk/zl@369