Since syntax macros are just macros in a special namespace, remove them as primitives and make them macros.  Unfortunately this exposed a problem with how syntax lists are handled.  Thus...
No longer automatically splice in syntax lists.  They will in most
cases get flatted automatically as the syntax is parsed, and the few
cases this is not the case the splicing can be made explicit by
prefixing the mid with a "@".  To make this work I had to add flatting
in two places.

svn-id: trunk/zl@373