Get ride of TopLevelNode and instead chain definitions together in the order they where defined.  To do this all Declaration are now Stmt.
This changed the order variables are compiled in modules so that the
definition might be after a function they are used in.  To deal with
this var. forward declarations are first compiled like function
decl. are.

This in turn exposed the fact that I didn't handle variables being
declared more than once.  Fixed.

svn-id: trunk/zl@376