And necessary support to allow fixing the size of the vtable by basically specifying the syntax form used for creating the class for the vtable, see test101.c.
This exposed some hygiene issues, mainly that marks were being ignored
on syntax forms.  This was corrected but this also created a problem
for the class macro which also creates classes.  Thus the
env. snapshot is now taken when the proc. macro is declared so the
"class" macro is part of the captured env.

Also added the ability to add flags to syntax lists.

svn-id: trunk/zl@397