Bug fixes, mostly related to the case when this pointer adjustment is needed when up or down casting.
When freeing an object, the this pointer must be adjusted to the most
derived type before calling "free".  To accomplish this the
"_destructor" function now returns a void pointer to this.  Which, if
virtual inheritance is involved should be virtual.  The case when it
is not virtual is not handled right now, and I'm not sure if it's
legal C++.

Do not make the _vptr macro a lvalue.  Since this involves a cast
setting the casted value is incorrect.  Instead, create a macro
_set_vptr, to use instead.

In C (and thus zls for now) assigning one array to another is illegal.
Thus use memcpy in init-assign.

Other minor fixes and enhancements.

svn-id: trunk/zl@407