Make the main class macro a ".class" that is a simplified version of a class.  Mainly to use virtual methods.  With this class-new_abi.zl no longer needs to me a patch since it will now simply override the necessary methods of the main class macro.
Also redo how the prelude is compiled so it is no longer a special
case.  "class-simple.zl" now gets compiled separate into its own macro
library.

Add support for "including" a file which unlike import is really a
literal include.  Fix things up so that include (and imported and the
like) files are searched for in the same directly the file including
it is at.  To do this it is necessary for all file names to become
absolute paths.

svn-id: trunk/zl@420