Misc. bug fixes and minor enhancements and add support example ABI that passes the this parameter in a global variable (to simulate a reg.)
Refactor the class macro API to call methods for creating method calls
and the like.

Recognize the "add_prop" directive separately in the class macro, to
avoid putting in the front of the module.

Bug fixes so that the prelude can be compiled as part of the program
rather than separately.  This involved some changes in class-simple.zl.


Remove unused code in main.cpp

Properly handle lowering of eblock and temporaries in destructor code.

Make function parameters AutoVar for now so that that if a local
var. shadows it both will have unique names.

Finally add new ABI is class-reg_this.* in test/.  No test cases, I
just forced it to be the default ABI and things worked.  See commented
out code in main.cpp for how.

svn-id: trunk/zl@424