Two closely related improvements to the peg parser.
(1) Do not create syntax objects in NamedProd unless 1) an explicit
capture is done or 2) the caller wants a syntax object (ie "SynBuilder
*" is non-NULL).  This saved around 5-10% time, and also cuts down on
memory a bit.

(2) Eliminate all unnecessary Captures.  That is only insert a Capture
object when it is really necessary to capture raw-text.  All pass
through captures are now eliminated.  This only gave a small 1-2%
performance improvement though.  Also add the helper struct ProdWrap
which will keep track of if we want a syntax object as a result and be
sure we only ask for it when it is needed.

Note, in the process of these improvements I eliminated the idea of an
implicit capture.  This changes the meaning of "A = B" so that A now
means to capture the resulting string from parsing B not the result of
B.  For the result "A = {B}" must be used.

svn-id: trunk/zl@486