Checkpoint.  Set things up to make future performance improvements easier.  These changes actually lead to worse performance on there own.
Add temporary option to disable error handling to get a handle on its
impact on performance (Major impact, will be improved latter).  

Add support for transient productions, i.e. named productions which
are not cached.  This requires a hint file which is not committed yet.
Also add support for dumping info which can be used to analysis the
performance of the peg parsing with regards to caching and thus create
the hint file by determining which productions are effectively
transient; the script to do so is, again, not committed yet.

Add support to determine the name of a captured syntax object (if
any) after it is fully parsed.  This allows common prefixed to be
folded is cases it wouldn't have been possible before.  For example:
  ID_W_NS = <::> {RAW_ID} "::" {ID_W_NS} / {RAW_ID};
  ID_W_NS = <<>> {RAW_ID} ("::" {ID_W_NS} = <::>)?;
This transformation avoided duplications of named productions, thus
allowing them to be transient.

Manually fold common prefixes in grammar when possible, to allow for
more transient productions.

Other minor changes.

svn-id: trunk/zl@488