Determine the possible set of characters the first character can match for any production, this includes the matching the EOF.  Also keep track of the production acts like a predicate or is optional.
This 1) improved on r490 (Keep track of the first character needed to
match a production if there is one) by expanding the test for a set of
characters rather than just a single character.  More importantly it
allows optimizing ordered choice to using a lookup table to only try
productions which might match based on the first character.  When
there are only a few choices to try this may not be worth it, thus it
is only used when there are 3 or more choices and when the lookup
table will cut down on the number of choices by only one in the best
case.

The first character property is computed at the same time the
persistent property was computed.  The dependency algorithm used in
r489 was optimized based on the observation that the only dependency
of significance was the one first one called in the mutual recursion
chain.  Ie given.
  A calls B calls C calls A & B
C will fail due to the mutual recursive dep on A and B, but only the
fact that C calls A needs to be kept track of, the fact that it also
called B is irrelevant.

svn-id: trunk/zl@496