Redo how PEG results are cached.  Instead of every production having its own hash table, have one global hash table to store the cached results with the key being <prod,pos>.
In the process completely rework how persistent productions are
handled.  

In the past each production kept track of how far it read to in order
to know if it can be used again when reparsing a string.  This has a
fair amount of overhead involved (see commit r477) and due to
technicalities some results that could of been reused where not.  I
eliminated the need for this by using a separate cache when a
"reparse" prod. is parsed, this cache is then saved along with the
results of the parse in the resulting syntax object.  Technically this
is not the same thing, but in practice it works just as well.  In fact
it works slightly better since the results that couldn't be used
before due to technicalities now can.  This strategy also associated
the cached results with the relevant string (technically the reparse
syntax object) which means that when the string is no longer used the
cache will be garbage-collected.  Before, the the results has to be
kept around for the life of the program.

Creating a new cache for each reparse syntax objects meant that lots
of tiny caches where created.  Using a hash table for such small data
sets is an overkill and uses a lot of unnecessary memory.  To avoid
this problem I created a tiny_hash, which will store the results in a
linked list up until a predetermined threshold, after which a hash
table is then used.  The elements of the linked list are allocated
several elements at a time to improve locality.  By taking advantage
of the hash table's internal representation, no data needs to be
copied when converting the linked list a hash table.

In the past each NamedProd had its own cache, this made clearing
non-persistent productions between passes easy as you just clear the
caches of each non-persistent production.  Now with everything in one
cache this is no longer possible without going through the cache
one-by-one.  To avoid this each cached result now has a run-number
associated with it if the run-number of the result is not the same as
the current run-number the result is discarded (technically the hash
table entry is reused to store the new result).  This has very little
overhead.  The other alternative is to maintain a separate cache for
the non-persistent entries and clear that between each parse.
However, in practice keeping track of the run-number was more
efficient, probably because it allowed the cached cache to also to
used for the persistent results when reparsing a string; since a
separate cache was used for each parse the cache never became
over-crowded with unused entries.

All this lead up to a 2.8 speed increase, but for some reason also
increased memory usage by up to a factor of 2 (from a heap size of 6.5
MB to 12 MB in the prelude benchmark) .  The reason for the increase
in memory usage in unclear and likely has something to do with when
collections are made and the overhead of the additional little caches.
In any case it is not something that I am going to worry about.

svn-id: trunk/zl@506