Checkpoint.  Rough around the edges but everything works.  Also added some debugging support.
Go back to directly adding module to env, but return a handle to the
symbol for:
  - adding the symbol in the correct env.
  - completing the parse

If the object was parsed in a temp. env. than the handle contains
instructions for what needs to be done before adding to a real env.,
mainly instructions for adding top level symbols.  The handle also
contains instructions for how to complete the parse.  For a module
this is what to do for the second pass.

When pre-parsing or expanded if there a forward declaration of a class
(or module) the forward declaration must be part of the same env. the
class is parsed.  Ie given:
  class X {
    class Inner;
    Inner * inner;
    class Inner {
       ...
    };
  };
the entire class must be parsed or expanded in the same temporary
environment.  Otherwise, the inner var. might bind to the wrong
object.

svn-id: trunk/zl@537