Finished implementation of overloaded functions.  The rules to resolve overloaded functions are still primitive through; for example, foo(int) and foo(double) will still not work.
Support identifiers to overloaded symbols.  These ids don't have an
initial type.  When the id is resolved (using "resolve_to")to a
function type the appropriate symbol is found and a new id is returns.
For example in:
  void foo();
  void foo(int);
  void (*fp)() = foo;
In the last line, the id "foo" does not have a type since it could
either bind to foo() or foo(int).  However, when the id is resolved in
the assignment it binds to foo().

Support C linkage, which simply means not to mangle the symbol.  CT
callback functions implicitly have C linkage, and all functions when
compiled in ZLS mode also have C linkage.  Of course it can be
specified explicitly.

Move some tests to broken/ which I don't want to deal with right now.

svn-id: trunk/zl@552