Minor fixes to make overloading work better.
Remove CT_CONST as a separate type, as it wasn't used for anything any
it causes a compile time string to be a different type than "const char
*", which causes problems with overloaded.

Limit the scope of when an integer conversion is considered a
promotion as per. the C++ standard.

Remove some debug output that was left in on the last commit.

svn-id: trunk/zl@596