Checkpoint.  Support recognizing template syntax.
Template id parsing is context sensitive is C++, thus is order to know
if foo<4>(8) is a template call or "foo < 4 > 8" it is necessary to
know if "foo" is a template.  Thus, a special rule has been added to
the grammar to support this.  To support this it is now necessary to
pass in the (ast) environment to the peg parser.

Templated ids are declared using the syntax:
  template Foo

Also, avoid some global variables in the peg parser and instead pass
around a MatchEnv.

svn-id: trunk/zl@597