Checkpoint towards supporting template syntax.
In turns out that in order to properly support recognizing template
syntax it is necessary to fully parse each statement before even
looking at the next one.  Ie parsing:
  template Vector;
  class Vector<int> ...;
as:
  (@ (stmt ..) (stmt ..))
no longer cuts it.  Since the statement "template Vector" needs to be
parsed and the template id "Vector" added to the environment before
"class Vector<int> ..." is parsed.

Thus, statements are parsed one at a time.  To support this a new
function parse_prod is added which is like parse_str except that the
it supports only parsing part of the string, thus enabling parsing
"STMT" one at a time.

Since "{ ... }" where often parsed as a list of statements it is
necessary to delay this be introducing a new reparse form "@{}" which
is like "{}" except with the instruction that it is to be parsed as a
list of statements.  Add support to partly_expand for recognizing this
by reparsing the first STMT and leaving the rest to be parsed latter.

Other related changes, some of which might be removed latter as they
where dead ends.

Recognize "tid" when parsing decls so that the tid becomes a
parametrized type.  Add support for "UNKNOWN" parameter in parse_type
which means to default to a compile time integer unless wrapped by
".type".

Also cache some common syntax leafs such as "@" and "@{}" by creating
one global instance.

svn-id: trunk/zl@598