Implement primitive "template" support, finally.
Really all I am going is recognizing the template syntax and than
manging the name with the parameters while expanding. No support for
creating templates, just sort-of using them.

Since templates are now handed during the expanding phase, this commit
removes support for recognizing them directly in parse_decl and user
types.  Support for the "UNKNOWN" parameter type is left in, for now,
as it might be useful latter.

Also only support conversion of syntax objects to a SymbolName or
String using implicit conversion when the syntax is simple.  This was
to cache cases when I should of used expand_binding instead.  It also
caught a few other abuses, which where fixed.

svn-id: trunk/zl@601