Checkpoint.  Defining methods outside the class basically works.
If an optional variable does not match make sure that match variable
is undefined so in:
  m = match(m, ...(x y)...);
  m = match(m, ...(x @y)...);
if the optional variable y doesn't match m->var(syntax y) will return
null, rather than the value from the match in the first line.

Add support for generic pointer entities, to be able to store
arbitrary objects in a syntax object.  No type checking is done.  This
is used by the class macro to store itself as a propriety of the user
type.

New function lookup_fancy_symbol, which is used by the symbol
propriety code to be able to store and retrieve symbols for overloaded
functions, and also to take care of calling expand_id in the right
places.

And of course changes to the class macro, using the above changes, to
support defining methods outside the class.

svn-id: trunk/zl@605