Checkpoint, lots of independent fixes to pave the way to parsing constructors and destructor definitions which are outside of the class.
Bug fix for commit 602 so that compiling with "-d" works again.  This
involved a hack to fake the dependencies on malloc/free and
ct_malloc/ct_free.

Enhance sample_w_loc to optionally also print the raw syntax.

Add a semi-backtrace on error messages.  Basically when parsing 10 in:
  int main() {int x = 10;}
it will give something like
  ...
  While parsing exp: 10: 10
  While parsing decl: int x = 10: (var x 10)
  While parsing decl: int main() {...}: ...
With each line giving source location when it is known.  The idea is
to give some context when macro expansion can make it unclear where
the expression is coming from.  May just add too much noise...

When importing into an "open scope", ie a module, import the
symbols before all other symbols in the module so that imported
symbols will not shadow ones defined in the module.  However, if two
imports define the same symbol, the symbol in the second import will
still shadow the first.

Add the user_type name to the associated module so that X::X() will
parse correctly.

When replacing allow optimal mids to be used in replace, even if they
didn't match anything. Ie make:
  m = match(... (x y @ z) ...)
  replace (... (x y z) ...)
work as expected

Remove some dead code and comment out some leftover debug output.

Other minor changes.

svn-id: trunk/zl@607