Add support for C++ initializer list.
Also, throw an error message when a expanding a simple macro if
the wrong number of arguments or given.  This involved failing a
match if too many arguments are provided rather than ignoring the
excess ones.

Also fix a bug in replace_mid if a flag has more than one
argument.

Other minor fixes.

svn-id: trunk/zl@612