Checkpoint.
Correctly handle C++ "struct" tags.  That is:
  struct Y { struct X * x; }
is equivalent to:
  struct X; struct Y { struct X * x; }
NOT:
  struct Y { struct X; struct X * x; }
In order to do this, I had to: 1) Add a special flag to
declare_user_type to declare that the struct should be declared in the
outer scope. 2) Add a hack to parse_decl to add this flag when the
decl. is not by itself. 3) And finally, provide a mechanism for
declaring the symbol outside of the module; this is accomplished by
adding the symbol as normal to the module symbol table but flag it
with PUSH_TO_OUTER_SCOPE; the symbol table is than scanned and the end
of the module declaration and any symbols with this flag are added to
the outer scope.  Special rules are added to make sure this flag is
only honored when inside a module, and that things work correctly with
nested modules.

Also, correctly handle the case when the struct name is not given by
assigning it an "anon" name (and then relaying on marks to keep the
symbol unique) in the class macro.  To make this work, also fix a bug
where the outer-namespace part of the symbol name wasn't getting
append with a uniq number when required.

Remove some overloaded forms of error as there where more trouble than
they where worth and lead to incorrect matches.  Instead use the
symbol NO_LOC when no syntax object or other form of loc. information
is available.

Other minor changes and cleanups.

svn-id: trunk/zl@613