Checkpoint.
Various fixed on way to getting randprog to compile.  Most of
them deal with getting the "template" vector (not get committed)
macro to compile.

Also print out the source line as a comment for each statement
and var declarations to make debugging zl generated code a little
bit easier.  This including making an effort to preserve the
source info at the caller site when expanding macros, and do a
similar thing when reparsing, and parsing expressions and
statements.  Change the Reperse syntax object to no longer use
str_ as the string for the outer reparse info, so that it can
freely be changed without causing problems.

svn-id: trunk/zl@617