Add support for handling pointer to functions as member variables. This required special support for id only macros so that when given id(10) the "(10)" is not passed into the macro.  In the process also rewrote how overloading is handled a bit.
This is Part 1 of several unrelated changes I made several months ago
and never committed, and am now trying to figure out what I did.  I
picked out the changes as best I could and the above is a best guess
of what I did.  Any added testcases are new and test the functionally
of what I think the main purpose of the change is.

svn-id: trunk/zl@683