Compute compile-time value of switch statements during parsing as some const expressions can't be done by backend.
This is Part 2 of several unrelated changes I made several months ago
and never committed, and am now trying to figure out what I did.  I
picked out the changes as best I could and the above is a best guess
of what I did.  Any added testcases are new and test the functionally
of what I think the main purpose of the change is.

svn-id: trunk/zl@685