Make extern "C++" chose more than just the mangler; it now also chooses the class implementation.  Alternate class implementation now register themselves in a similar manner that the mangler does. parse_class now gets the class implementation to use from the environment (via the new API function environ_get_abi_info).
Even though the ABI to use is part of the environment, it is separate
from the symbol table, which means that it is non-hygienic.  Ie, if a
macro generates a class the ABI used will be determined at the call
site rather than where the macro was defined.  I thought about making
it part of the symbol table but that could lead to weird results when
some parts of a macro result have marks and others don't.  For
example, suppose this was the output of a macro:

  class Base' {...};
  class Derived : public Base' {...};

than if the ABI choice was in part of the symbol table, Base' (with a
mark) and Derived (no mark) might get different ABI, which is probably
not what was wanted, and, in addition will probably not work since the
two ABI's are likely incompatible.

This change now activated the alternative class implementation for the
gcc ABI.  Right now all this does is emit two constructors and
destructors as required by the ABI.

svn-id: trunk/zl@694