Make the delete and new operator a function of the class or the active ABI for primitive types.
To do this I now create a special module for holding symbols specific
to the ABI.  When new or delete is called I look for the new and
delete macro in the ABI module if it exists, otherwise I use the
default implementation.  When determining the ABI to use for a class
first the user type is consulted, then the environment.

With this change the correctly mangled new and delete function are
called when using the gcc ABI.  For now, using the gcc ABI requires
linking with libstdc++.

svn-id: trunk/zl@696